backend/requirements.txt
# Core API and Web Framework
anthropic>=0.35.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Scientific Computing and Visualization
matplotlib>=3.8.0
numpy>=1.24.0
scipy>=1.11.0
sympy>=1.12
pandas>=2.1.0
seaborn>=0.13.0
imageio>=2.34.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0 